add missing promises in node:dns
